requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
torch==2.2.1
faiss-cpu
numpy<2

